k8s-runner: the first kubernetes runner
This runner is experimental.

This piggybacks on the api in https://github.com/tinyci/k8s-api to
create a chain of runner<->k8s bridging through use of CRD management.
The CRD maintains the state of the pod and other resources and
communicates the overall status back to the runner.

The runner can also now parallelize runs across the cluster, with a
`max_concurrency` argument that takes an integer of what to run.
Eventually, it would be nice to support smarter resource management.

This, again, is experimental.

Signed-off-by: Erik Hollensbe <github@hollensbe.org>