overlay_tempdir argument for overlay-runner
This argument controles the base path of the overlay temporary
directories created during the testing process.

The container needs to be able to mount overlayfs which has to be on a
"real" filesystem, so bind-mounting this path to the host + setting this
value will allow overlay-runner to run in your container with reasonable
performance. Neat!

Signed-off-by: Erik Hollensbe <github@hollensbe.org>